requirements.txt
# Core ML/DL frameworks
tensorflow>=2.16.0
keras>=3.0.0

# Data processing
numpy==1.24.3
pandas==2.1.0
scikit-learn==1.3.0
opencv-python==4.8.0.74
Pillow==10.0.0
albumentations==1.3.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# ML utilities
scikit-image==0.21.0
imbalanced-learn==0.11.0

# Experiment tracking
mlflow==2.7.1
tensorboard==2.15.0

# API/Deployment
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
joblib==1.3.2
requests==2.31.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Development
jupyter==1.0.0
ipykernel==6.25.1
black==23.7.0
flake8==6.1.0
